@@ -13,7 +13,6 @@ class CustomSources(BaseSettings):
             dotenv_settings: PydanticBaseSettingsSource,
             file_secret_settings: PydanticBaseSettingsSource,
         ) -> tuple[PydanticBaseSettingsSource, ...]:
-            print(source)
             return init_settings, env_settings, source(settings_cls), file_secret_settings
 
         cls.settings_customise_sources = classmethod(func)
@@ -27,7 +26,7 @@ class TomlSettings(CustomSources, source=TomlConfigSettingsSource):
 
 class YamlSettings(CustomSources, source=YamlConfigSettingsSource):
     """
-    可以先写好yamle，使用 datamodel-code-generator 生产模型
+    可以先写好yaml，使用 datamodel-code-generator 生成模型
     """
 
     model_config = SettingsConfigDict(yaml_file="conf.yaml", extra="ignore")
