@@ -27,7 +27,7 @@ class BaseSerializer(BaseModel):
 
     # 验证字段时实例还没有构造，不使用self
     # mode="before" 验证原始输入
-    # mode="after" 验证转换为定义字段后的值, 默认值
+    # mode="after" 默认值, 验证转换为定义字段后的值
     @field_validator("created_at", "updated_at", mode="before")
     @classmethod
     def transform_time(cls, v) -> datetime:
