@@ -4,40 +4,35 @@ from sqlalchemy.orm import *
 
 from conf import settings
 
-# 修改默认方言为psycopg
-# from sqlalchemy.dialects import postgresql as pg
-# pg.dialect = pg.base.dialect = pg.psycopg.dialect
-
 # os.environ["PGTZ"] = "utc"
 # psycopg: connect_args={"options": "-c TimeZone=Asia/Tokyo"}
-# url = URL.create()
+
+# URL.create()
 # default pool_size=5
-db = create_engine(settings.db, echo=False, pool_size=10)
-Session = sessionmaker(bind=db)
+engine = create_engine(settings.db, echo=False, pool_size=10)
+SessionMaker = sessionmaker(bind=engine, expire_on_commit=False)
 
-
-def async_session():
-    async_db = create_async_engine(settings.db_asyncpg, echo=False, pool_size=10)
-    async_session = async_sessionmaker(bind=async_db)
-    return async_session
+async_engine = create_async_engine(settings.db_asyncpg, echo=False, pool_size=10)
+AsyncSessionMaker = async_sessionmaker(bind=async_engine, expire_on_commit=False)
 
 
 def usage_a():
-    with Session() as session:
+    with SessionMaker() as session:  # type: Session
         session.execute(...)
         session.commit()
+    # outer context calls session.close()
 
 
 def usage_b():
-    # inner context calls session.commit(), if there were no exceptions
-    # outer context calls session.close()
-    with Session() as session:
+    with SessionMaker() as session:
         with session.begin():
             session.execute(...)
+        # inner context calls session.commit(), if there were no exceptions
+    # outer context calls session.close()
 
 
 def usage_c():
     # include begin()/commit()/rollback()
-    # commits the transaction, closes the session
-    with Session.begin() as session:
+    with SessionMaker.begin() as session:
         session.execute(...)
+    # commits the transaction, closes the session
